feat: craco를 이용한 절대경로 설정 및 디렉토리 구조 리팩토링
