+++ src/App.tsx
@@ -1,9 +1,9 @@
-import Router from 'Router';
+import Router from '@/components/router';
 import { ThemeProvider } from 'styled-components';
-import { lightTheme, darkTheme } from 'styles/theme';
-import GlobalStyle from 'styles/GlobalStyle';
+import { lightTheme, darkTheme } from '@/styles/theme';
+import GlobalStyle from '@/styles/GlobalStyle';
 import { useRecoilValue } from 'recoil';
-import { themeModeAtom } from 'stores/atoms';
+import { themeModeAtom } from '@/stores/atoms';
 
 function App() {
     const isDarkMode = useRecoilValue(themeModeAtom);
