+++ src/index.tsx
@@ -1,7 +1,7 @@
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from '@/App';
 import { RecoilRoot } from 'recoil';
-import reportWebVitals from './reportWebVitals';
+// import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(
     document.getElementById('root') as HTMLElement,
