+++ src/pages/CardBook.tsx
@@ -1,4 +1,4 @@
-import Layout from 'components/common/Layout';
+import Layout from '@/components/common/Layout';
 
 // 아티스트 카드 도감 페이지
 function CardBook() {
