+++ src/pages/Register.tsx
@@ -1,0 +1,3 @@
+export default function Register() {
+    return <div>회원가입 페이지</div>;
+}
