+++ tsconfig.json
@@ -1,6 +1,5 @@
 {
     "compilerOptions": {
-        "baseUrl": "src/",
         "target": "es5",
         "lib": ["dom", "dom.iterable", "esnext"],
         "allowJs": true,
@@ -17,5 +16,7 @@
         "noEmit": true,
         "jsx": "react-jsx"
     },
-    "include": ["src"]
+    "extends": "./tsconfig.paths.json",
+    "include": ["src", "src/**/*"],
+    "exclude": ["node_modules", "build"]
 }
